Add prep7.
Note: if it is slow to run vagrant commands, this is because vagrant is
slow to process a complex Vagrantfile that has many machines defined. If
you destroy or halt the unused vms, and then comment them out, since
there are fewer to loop through, the vagrant operations should go
faster!